Add iOS Support
The generated .framework was added to the kubenav-plugin, so we can add
iOS support via the plugin.

Attention: The Bug mentioned in #1 was already fixed, but not released
so we manually adjusted the code to build the iOS version. When a new
release of Capacitor is available we need to update the used version.