Add alternative cluster option for request method
The request method for the context doesn't have an option to pass in a
alternative cluster then the currently selected cluster. This leads to
problems with the cluster status in the cluster overview. There is the
status of the active cluster shown for every cluster.

Now it's possible to pass an alternative cluster to the request method,
which then is used instead of the active cluster.