Improve error handling for failed API requests
The error handling, if a resource could not be loaded was improved. If
the clusters object or the cluster string is not defined we set the
values to undefined to improve the code. If an API request fails we are
showing an error message on the page instead of an alert modal.