Add indicator for clickable elements
Some items were not indicated that they are clickable. Therefor we added
the button attribute to these components. Now they have an indicator
when they are clickable.

For several modals the close button was not marked as clickable. This
was also fixed by adding an IonButton component to the IonButtons group.