Improve logs handling
By default only the last 1000 log lines are loaded, because the complete
log results in very high traffic. This could be problematic on mobile.
The complete log can be loaded via a new menu button in the logs view.

It's now possible to load the previous log of a container. This should
improve the debugging capabilities of kubenav. Like for the current log
there are two options: Load the last 1000 log lines or load the complete
log.

To refresh the log, the corresponding menu item must be pressed again.
The refresh button is gone in favour for options menu.