Improve log access on desktop
To access the logs on desktop a button was added to the container items.
Now it's not necessary anymore to swipe on the container name to access
the logs on desktop.

We had to add e.stopPropagation() to every button which can be accessed
in the logs view, to avoid some side effects with the containers modal.
The modal was opened without this function.