Add support for GKE via OIDC
GKE clusters can now be added via OIDC. Therefor a user has to
authenticate himself via Google OAuth2. The returned code is exchanged
for a refresh and access token. The access token is used to get a list
of projects where the user is a member of. In the final step all GKE
clusters for all projects are retrieved. A user can then select all
clusters he wants to add to the app.

If a request to a Kubernetes cluster at GKE is made, the current access
token of the user is checked that it is not expired. If the access token
is expired a new access token is requested with the refresh token.

Note: If the retrieved cluster data contains a username and password or
a client certificate and key, these credentials are used instead off the
access token.

Note: The authentication flow is documented under the following URL:
  https://developers.google.com/identity/protocols/OAuth2InstalledApp