Add support for EKS
To import Kubernetes clusters from EKS a user have to provide an access
key id, a secret key and a region. These credentials are used to retrive
all available EKS clusters from the AWS API and to generate a token for
the Kubernetes API requests.

Only one access key id and secret key pair per region is allowed, but it
is possible to use different credentials per region.

Note: Currently for every Kubernetes API requests a new token is
requested from the AWS API. These tokens expires after 15 minutes. These
can be improved to only request a new token if the old one is expired.
Since the tokens are only valid for the specified cluster name, they
must be regenerated for every change of the cluster context.