Use kubeconfig file on desktop
Instead of manual adding clusters to kubenav on desktop, the kubeconfig
file of the user is used now.

At the start of the server component, the local kubeconfig file is
parsed. To use the clusters from the kubeconfig the server implements
two new endpoints cluster and clusters. These endpoints returning the
current context and all clusters from the kubeconfig file.

While the kubeconfig file is used now, it is not possible anymore to add
clusters manual to kubenav. But all cloud providers are supported on
desktop now.

When kubenav is started the current context and namespace is
automatically set as initial selected cluster.