Adjust electron to work with kubeconfig
Since we are using the local kubeconfig file now, we had to adjust the
electron logic to show and switch clusters.

Now we are saving the selected cluster in localStorage and check if the
name form localStorage exists in the loaded clusters. When the name
exists this cluster will be used instead of the loaded active cluster
from the current context.