Improve pod and container view
Show the resources and the current usage for each pod in the list view.
Show resource, usage, status, restarts and ready status for each
container on the pods detail page.

Since the container modal has a width of 90% on larger screen, we show
some of the cards next to each other. The modal width and height was
changed in PR #22.