Fix logs modal for valid json
If the returned logs from the Kubernetes API server are a valid JSON
object we are parsing the log and returning an object instead of a
string.

This had break the modal and the app crashes. This is now fixed, by
checking if the returned data is a string. If not we are reverting the
parsing befor passing the data to the Editor component.