Do not open container modal when clicking on logs
Fix a bug where the container modal was open when log lines in the
editor component where selected. The bug occures because we didn't pass
the e.stopPropagation() function to the Editor component. This function
is needed to add the logs button to the container item on desktop.

Instead of adding it to the Editor component we moved the Logs modal out
of the IonItem component for the container.