Add searchbar for list views
With a lots of resources in the list view, endless scrolling can be a
little bit annoying. To solve this problem we add a searchbar to the
list views.

Only the name of the resource is included in the search. It is not
possible to search for resources by other properties.