Add support for OIDC
It is now possible to configure an OIDC provider to access the
Kubernetes API server. You can also provide an existing refresh token
for the OIDC provider, which then will be used to retrieve the id token
to access the Kubernetes API.

If you configure your OIDC provider you have to choose
https://kubenav.io/oidc.html as redirect URL. On this page you will see
an "Open App" button to go to the app after a successfull
authentication. This is necessary, because the most providers doesn't
allow the redirect directly to the app.