Check if authProvider property exists on cluster
The authProvider was introduced in a later version of kubenav. Therefor
it is possible that the property doesn't exists on old cluster
configurations. This results in an error, which can only be fixed by
readding the cluster.

This commit fixes the check for the authProvider property and applies
the OIDC configuration only if the property exists.