Add version check
Add a version check which compares the currently running version of
kubenav with the latest version tag from the GitHub releases. When the
tags are different a menu item "Update Available" is added. When the
item is clicked the GitHub release pages is opened in the browser.

Note: This only works for the production releases. If a beta release is
used the "Update Available" item is always shown in the menu.