Add support to stream logs
Add a new options to the logs popover to stream log files. If the
streaming option is selected a request is made to generate a session id.
This session id can be used to retrieve the logs via Server Sent Events.

Unlike the other options, the stream logs option uses xterm to show the
logs (for the other options we use Ace), because the editor would be
rerendered when retrieving a log line.