Add search function for terminal and logs
It is now possible to search throw the terminal and logs. The searchbar
can be open via the popover in the upper right corner in the modal.

Increase the scrollback buffer to 10000 lines.

Fix the positioning of the logs popover on modal. The modal wasn't
placed next to the button, instead it was centered at the screen.