Improve rendering
Optimize rerendering of components by using React.memo. Some renderings
which were caused by React Router are unnecessary. By using React.meno
these rerenderings are avoided.

Introduce useAsyncFn to improve the data fetching in the components.
This needs some more refactoring in components which were not touched.
The introduction of useAsyncFn fixes also a bug where the wrong page was
displayed. This could happen when the page was changed before the API
call was finished.