Rework status indication
Use a status function instead of a status variable to determine the
status of a resource.

Distinguish between warning and danger. The status should be danger,
when required status indicator is zero or undefined. When this isn't the
case and the success probe failes the status is set to warning.