Restart server for API requests on mobile
After a period of inactivity the background task for the server is
stopped on mobile.

To check the health status of the server a new API endpoint
'/api/health' was added. This endpoint is requested befor each request
to the server. If the status code is not OK, the server is started again
befor the original request is made.