Improve context switch
Improve the switch between contexts/clusters by adding a new select box
to the top of the menu. The select box always shows the current context
and by selecting an action sheet is opened to change the context.

After selecting a context the user is redirected to the home page.

The user haven't to go to the clusters page now, to switch the
context/cluster.