Update npm packages
Update all npm packages to there latest version, except eslint. eslint 7
is currently not supported by react-scripts.

Adjust eslint rules for the latest version and the resulting eslint
errors and warnings.

Remove the scroll down button for the editor, which was only used by the
logs. The logs are using xterm.js, so that the button isn't required
anymore.