Fix dark/light mode support
It was not explicit checked if the darkMode settings exists or it was
false. Therefor the disabled dark mode wasn't used on a system where the
dark mode was enabled by default.

Now it is checked if the darkMode property exists in the settings and
only if it doesn't exists the default value will be used.