Add support for PDBs
The config and storage section contains a new entry for
PodDisruptionBudgets.

The status of a PDB is dependens on the number of expected Pods and the
current/desired healthiness Pods. If the number of expected Pods is 0,
the status will be warning, because this means the PDB selector doesn't
match any Pods. If the current number of healthy Pods is lower then the
number of desired healthy Pods the status will be danger.