Support selection and copy in terminal
Because of the limitation of xterm.js on mobile it wasn't possible to
select and copy text from the terminal. Therefore a new entry was added
to the terminal menu, which enables the selection and copying text.

The text can be selected via a range slider and copied via button. Maybe
the range selection must be changed later, because it can be difficult
to select the correct lines, when the number of displayed lines is very
large.