Fix custom key event handling and change API route
Fix a bug in the custom key event handling, whereby the shell for Pods
stopped working. No key events were passed to the websocket connection.

Change the API route for the exec handling, to include "exec" in the
path. This is the preparation to add a new handler for the SSH
connection to nodes.