Fix pod status
The pod status wasn't shown correctly, therefor the "Status" field is
replaced by "Phase" and if available "Reason".

If the phase of the pod is "Running" or "Succeeded" the status is
success. If the phase is "Unknown" the status is warning and for
"Pending" and "Failed" the status is danger.