Add migration logic for old cluster configuration
Add logic to migrate the configuration of old clusters to the new
format. This is required so that the user doesn't have to readd all his
configured clusters.

The logic will be executed during the first start with the new version,
after that all existing localStorage items will be removed and the
migration logic is disabled.