Improve error handling for requests
Return status 400 (bad request) instead of 500 (internal server error)
for requests were an error happend. This was changed because a request
can also fail if the user hasn't enough permissions to view or edit an
resource. When kubenav is running in a cluster behind an Ingress the
high rate 500 can trigger an alert, so instead we are returning 400.

Now we are also returning the detailed error message for a failed
request, to give the user better feedback on what was going wrong.