Fix namespace switching bug for Electron
It could happen that the Electron app stopped working, when a user
switched fast between different namespaces. This bug was caused by the
event listener in the Menu component, which handles the navigation from
the Electron menu.

When the namespace was switched a new listener was initialized, which
results in to many pending requests.