Use cobra for command and flag parsing
Instead of Go's flag package we are now using cobra. A new 'version'
command is included which prints all important version information for
kubenav. The names of all flags were also adjusted to use a more
consistent naming.

This is the preperation for some upcoming changes were kubenav make more
use of flag, especially when kubenav is deployed in a cluster.