Fix permissions table
Fix a bug in the permissions table, were the permissions was showed
wrong when the allowed verbs includes '*'. Instead of a green checkmark
a red cross was shown, because it wasn't checked if the verbs list
contains the asteriks character.

This fixes the view for Pods, Service Accounts, Roles and Cluster Roles.