Add support for Port Forwarding
When the user selects a port of a Pod, Container or Service he see an
option to start a port forwarding session to the selected port.

If the user selects a service, a list of all Pods is requested which are
matching the selector and the first Pod in the list is used for the port
forwarding.

The user can set a local port which should be used or he can choose 0
and a random port is selected.