Improve naming of terminal tabs
Currently it is not that easy to close the correct terminal tab, when
there are multiple tabs with the same container name. Therefor we are
adding the pod name to the tab name.

For multi-line rendering we have to remove the IonLabel from the popover
where the tab can be closed. We also have to disable the max-width of a
tab.