Add Prometheus plugin
Add new Prometheus plugin, which is used to render Prometheus data via
Recharts. The requests to a Prometheus Pod are handled via Port
Forwarding.

In this first version kubenav supports Memory, CPU, Network and
Container Restarts graphs for Pods and Containers. For PVCs it supports
the current Volume space and inodes usage.