Add status for PVs and PVCs
The list view for PVs and PVCs are now showing a status.

If the Persistent Volume is in the phase released or available the
status is set to warning. When the phase is failed the status is set to
error and for bound to success.

If the Persistent Volume Claim is in the phase bound the status is set
to success. For all other cases the status is set to warning.