Clean up frontend
Clean up the frontend and remove less important features from the
details page (e.g. Volumes, Affinities, Tolerations, ...). Instead of
these components each details page contains a new menu point, which
allows a user to view the YAML of an resource.

We also renamed the SERVER and IS_SERVER variable to INCLUSTER_URL and
IS_INCLUSTER to have a consistent naming for this feature with the
backend code.

The cluster selection in the menu is now disabled when kubenav runs
inside a cluster. Since the name of the cluster is not visible to the
user anymore, also the incluster.name flag was removed.

The settings were rearanged, the connection button for SSH connections
were removed (instead a click on the address in the status is enough),
unused components and functions were removed, a new hook to get the
screen width were introduced (this will be used later to improve the
layout) and a lot more small adjustments.