Support Prometheus plugin with incluster mode
This commit adds support for the Prometheus plugin, when kubenav is
running with the incluster mode. For this we are loading the settings
for the plugin which can be set via command-line flags and adjusting the
logic for the Prometheus plugin, so that the Prometheus pod isn't
requested. This isn't necessary, because with the incluster mode we are
not using port forwarding for the plugin, instead we are using the
cluster address of Prometheus.