Rework Prometheus Plugin
Instead of the single charts we are now rendering all charts within a
single dashboard. This means the time range of all charts can be
adjusted within a single time range selector, we can use variables (e.g.
the container of a pod) and we only have to create one port forwarding
session to retrieve all data from Prometheus.

To speed up the query time, we are processing all charts in parallel on
the backend side.