Support automatic theme changing
This adds a new setting named theme. The user can choose between system,
dark and light. If system is choosen, the theme of the app automatic
changes, when the system theme is changed.

The old darm mode setting is removed.

This doesn't fix the problem that a theme change (manual and automatic)
isn't applied to charts, terminals and editor windows.