Use infinite scrolling for list pages
To avoid large responses which can cause problems within the app, we are
now using infinite scrolling for all list pages. To make use of infinite
scrolling we have to replace the useQuery function with the
useInfiniteQuery function from React Query.

The list which are shown inside a details pages (e.g. Pods that belong
to a Deployment) the number of items is limited to 100.