Refetch data for list and details pages
All data for the list and details pages is now automatically refetched
after 5 minutes. This interval can be configured by the user in the
general settings. The user can choose an interval between 30 seconds and
10 minutes.

The setting is not applied for all queries, because this can cause
problems for some components. For example, when we use this in the
context component, this will cause a rerendering of the page every time
the data is fetched.