Readd container details
During the clean up of the frontend we removed the container details, to
decide which parts are relevant and how to present them to a user.

Now we are readding the container details. The details modal contains
the container configuration, status, environment variables, volume
mounts, liveness, readiness and startup probe of a container. We removed
the resources and security context from the details modal. The resources
are already shown in the pod details view and the security context can
be accessed via the YAML view.