Close item options and popovers after selection
To improve the usability of kubenav, the shown item options and popovers
are now closed when the user selects an action. Until now they were
still shown in the background, so that a user had to close them
manually.

This is relevant for the details popup the resource options in list
views (YAML view and delete action) and the container items at the pod
details page.