Prepare support for bookmarks
To introduce the new bookmarks option, we have to remove the refresh
button from the list and details page and move this option to the
details menu. Since there wasn't a details menu for the list page this
was also added.

The new bookmark button will be available via the details menu.

We also adjusted the determination when to use the mobile layout.
Therefor we are checking the platform for 'mobile' to have a more native
feeling when the web version of kubenav is opened on a mobile device.