Add support for bookmarks
This commit adds support for bookmarks, each bookmark contains a title,
the url of the current page and the namespace. The namespace is needed
to create a bookmark, so we can get all Pods from a specific namespace.

The bookmarks can be created/deleted on each page by selecting the
bookmark item from the details menu. All bookmarks are shown on a new
bookmarks page, from where the user can go to the bookmarked page or he
can delete a bookmark.