Refresh lists in details page
If a deployment is scaled down to zero and the deployment page is
refreshed it can happen that there are still pods visible in the pods
list. This can happen because we only refresh the list when the
deployment changes, but some pods can still be terminating.

Since we do not want to call the refetch function of the list, we just
use the same refresh interval for the list as for the details page.