Add cluster overview page
Until now it was not possible to get an overview of the used resources
in a cluster. Therefore we have added a cluster overview page, with the
metrics for CPU, Memory and Pods and all warning Events.

The overview page replaces the old HomePage component as start page for
kubenav. Later we can add some more details for workloads and the
Prometheus metrics.