Parse Kubeconfig containing OIDC Provider
Until now it wasn't possible to add a Kubeconfig which contains clusters
using an OIDC provider. kubenav rejected these config with an error
message "Invalid kubeconfig".

Now it is also possible to paste/select a Kubeconfig file containing an
OIDC provider. The auth context is now used by kubenav and the cluster
is added in the same way, like a cluster using token/certificate based
authentication.