Add support for creating manifests
It is now possible to create new manifest files, within the details menu
in the list view of each component. This feature is available for each
resouce kind, except ComponentStatuses. CRDs, Events and Nodes.

To implement this feature we had to add the apiVersion and kind to each
page, which we may also should use for the generation of the request
URLs in the future.

The user have to enter the name for each resource. For namespaced
resources we are also requesting the namespace. For some other resources
like Deployments, StatefulSets, etc. we are also showing some additional
fields in the form (e.g. replicas and image). When a user clicks on the
Next button the complete yaml file is shown where the user can add
additional properties. With a click on the Create button the yaml file
is applied in the Kubernetes cluster.