Adjust templates to create resources
This commit fixes the indentation of the created templates and adds
support for CronJobs and Pods.

We are also changing the logic how the form values are handled, to have
only one variable for all form values now. The namespace is also
prefilled with the currently selected namespace.

This also fixes a bug with the Editor component, where the Keyboard
logic is now only used for mobile, since the methods are not implemented
for the web and desktop version.