Add status of Helm release to list view
We are now displaying the status of a Helm release in the list view,
like it is done for Pods, Deployments, etc. The status is displayed as
danger when the Helm release failed. When the Helm release status is
deployed, superseded or uninstalled the status is displayed as success.
For all other values we are showing a warning.

We are also fixing a bug, where the maximum callstack size was exceeded
for large Helm releases, by using the solution from
https://stackoverflow.com/a/60782610/4104109.