Improve Helm Plugin
We are using the offical Helm icon now in the sidebar and for error
messages. It is also possible to show all version of a Helm release in
the list view. This setting is saved in the localStorage so that we
always use the users last setting. Last but not least we are showing the
History of a Helm release in the details view of a Helm release.