Disable link for current Helm release in history
When the link of the same Helm release was clicked in the history, it
looks like nothing happens to the user, but the redirect to the same
page is added to the history, so that it also looks like nothing is
happend when the user clicks the back button.

To fix this weird behavior we disable the link in the history, for the
currently viewed Helm chart.

Fixes #243.