Add support for AWS SSO
This commit adds support for AWS SSO, were the user can add EKS clusters
by using the AWS SSO flow. For that the user have to provide a start URL
for the SSO flow, his account id, a role name and the AWS region. When
the user has provided all of these details a new client will be created
which the user has to verify. As last step he can use the new client to
import all of his EKS clusters.

Currently the returned access token is valid for around 8 hours, after
that time the user have to redo the complete authentication flow. For
that we should add a reauthentication method to the edit cluster page,
like it was done for Google and OIDC.