Improve AWS SSO Support
It is now possible to edit the AWS SSO credentials for a cluster and to
re-authenticate when the access token is expired. We are also supporting
AWS SSO when the SSO config and the Kubernetes cluster are using
different regions, for this the user must now provide the SSO region and
the cluster region during the setup.