Add setting for the query limit
Until now th limit for some queries were restricted to 100 items, which
could cause problems that some resources not shown. To allow the user to
query more items, when needed, we are adding a new limit setting.

The upper limit is set to 1000 items, each user must ensure that a to
high limit has a negative impact on the performance of the app.