Improve event handling
To improve the handling of events we are adding the namespace of the
event in the overview page, showing the full event message in the events
list, adding the link to the involved object in the event details and
allow filtering for the event type, reason and message in the events
list via the search input.

We are also adding a button to load more items, which is needed, when
the user type sth. in the search box. Since it can happen that the
filter reduce the number of items to not fill the screen, the load more
function isn't triggered via infinite scrolling. The button can be now
used to load more items in a filtered view.