Add timeout for terminal
We have to add a one second timeout after a new terminal and SSH session
is initialized, to make the terminal fail less often. This is caused by
the time which we need to initialize the session in the backend and so
the frontend requested the connection, when the terminal isn't ready for
usage.