Add support for DigitalOcean
It is now possible to import Kubernetes cluster from DigitalOcean. For
that a user must provide a personal access token, which is then used to
load a list of Kubernetes clusters and the Kubeconfig for each cluster.

If the access token for the Kubernetes cluster is expired, the user can
update the Kubeconfig in the edit modal for that cluster.