Add authentication support for mobile app
To make kubenav more secure on mobile, we are adding support for
authentication. The user can enable authentication in the general
settings.

This feature requires Face ID or Touch ID on iOS and fingerprint or face
recognition on Android. When authentication is enable a user will not be
able to use kubenav without it.

The implementation of this feature uses the Cordova Plugin Fingerprint
All-In-One:
- https://github.com/NiklasMerz/cordova-plugin-fingerprint-aio
- https://ionicframework.com/docs/native/fingerprint-aio