Fix re-authentication when adding a new cluster
When a new cluster was added the user was forced to re-authenticate,
when authentication was enabled, because we were using some ugly
window.location function for redirecting. Now we are properly using the
underlying React Router to handle the navigation, after the user
selected an option to add a new cluster.