Improve Docker image and add logging
We improve the build process of the Docker image by adding a
.dockerignore file, which reduces the size of the build context. We also
reworked the build images, so that we are now caching more layers in the
build process.

We also added a bit more logging, to make it easier to debug issues for
the desktop and server version of kubenav.