Disable close modal action for ESC key (terminal)
By default every modal can be closed with the ESC key. For the terminal
this can be problematic since we are also using the ESC as input
character for commands in the container. Therefore we disable the close
modal action for the ESC key within the terminal. The terminal can now
only be closed via the close button in the upper right corner.

Fixes #265.