Fix rederict for Google and OIDC
Currently the redirect within the app for Google and OIDC is broken,
because history.push doesn't pass the query parameters to the correct
page.

This is fixed by using the old way via window.location.replace. This
also means that a user must reauthenticate when he enables
authentication and is redirected to the app.