Fix authentication screen for larger screens
For larger screens (tablets), the Sign In button was hidden and the user
had to scroll to click the button. Now the button should be visible for
all display sizes.

For that we add a new hook useWindowHeight, which returns the current
height of the window. If the window height is to small to display the
icon and the Sign In button we hide the icon.

We also rename the Sign In button to Unlock.