Add advanced search functions
Until now it was only possible to search by the name of an item. The
search function is extended, so that it is possible to search every
field of a resource.

A user can specify the path to an field of an object, by starting his
entered search term with a dollar sign followed by a point and the path
to the field. Then he has to enter a colon and the value which the field
should contain.

Examples:

- To get all services of type load balancer the user can use the
  following search term: $.spec.type: loadbalancer
- To get all warning events a user can use the following:
  $.type: warning