Add support for dashboards
It is now possible to create custom dashboards. The dashboards are
loaded from ConfigMaps and displayed in the new Promtheus Dashboards
page.

If a resource contains the "kubenav.io/dashboards" annotation, the list
of dashboards is displayed on the details page of this resource. It is
also possible to pass for example the name of the resource as variable
value to the dashboard.

The unit for dashboards is now displayed at the y axis and not in the
title. Also a new chart type "singlestat" was added. This chart is
similar to the singlestat panels from Grafana.