Add Elasticsearch plugin
This commit adds a new plugin for Elasticsearch. It uses the same logic
as the Prometheus plugin. This means the mobile and desktop version is
using port forwarding and for the web version the plugin is configured
via environment variables.

The plugin allows a user to query a configured Elasticsearch instance
directly within kubenav. A user can exute query manually or can add an
annotation "kubenav.io/elasticsearch-queries" to a resources with
predefined queries for Elasticsearch.