Improve plugin security
We are not passing the username and password for Elasticsearch to the
client in the web version. We are also adding support for basic
authentication for Prometheus.

Fix a small bug, where the list of fields was empty in some cases.

Improve style for table layour of Elasticsearch search results.