Fix port forwarding
Fix port forwarding for the mobile version of kubenav. The port
forwarding does not require the full url since the #270.

Fix the error handling, when a port forwarding request fails. The
request never returned a response, when an error occurred. These errors
are now handled and shown to the user.