Add Jaeger Plugin
The new Jaeger plugin allows users to view traces from Jaeger within
kubenav. The configuration of the Jaeger plugin is similar to the
configuration of the existing Prometheus and Elasticsearch plugin.

Traces can be added to a resource in the cluster via the
kubenav.io/jaeger-traces annotation. The traces are defined via an JSON
array, where the title and the service name for Jaeger is required. All
other fields are optional.