Fix scrolling for info page
On iOS it could happen that a user can not scroll in the info page and
so the links are not reachable. Our current assumption is, that the
problem was caused be the displayed paragraphs, so the paragraphs are
removed for now.

We are also removing the displayed image for larger screens, because on
large screens only the image was visible and nothing else.