Improve default metrics
Instead of container_memory_usage_bytes we are now using
container_memory_usage_bytes for the memory metrics now. The advatange
of this that the graphs will be consistent with the metrics retrieved by
the metrics API.

For the node and cluster metrics we are also switching to this metric.
So that the results for the graphs are the same as reported by kubectl
top nodes.

We also have adjusted the metrics for CPU usage and the Network I/O.