Remove donation links in iOS version
The last submitted version was rejected by Apple, because we do not use
the In-App Purchase API for donations (3.1.1). Instead we used the
direct links for GitHub Sponsors and Paypal.

It looks like this isn't covered by point 3.2.1 (vii) of the App Store
Review Guidelines anymore. Unfortunately I didn't receive an
explanation from the review team, so that we have to remove the
donation links from the app.