Fix time range selection in Prometheus charts
When a user selected a new time range for the Prometheus charts, the new
data wasn't loaded. This should be fixed by adding the time range to the
query keys for react-query.

When the new data for the charts is loaded we also keep the previous
data, so that the UI isn't changed that often.