Fix kubeconfig selection for some Android devices
On some Android devices it looks like the selection of a Kubeconfig file
is still not working. To fix the issue we do not limit the accepted
files for the file input anymore on Android.

This shouldn't be a problem, because the filetype limit was mainly
introduced for iOS, to get through the review process.