Add live resource usage metrics
This commit adds live resource usage metrics for Pods. The metrics can
be viewed via the new "Live Metrics" action in the Pod details. When a
user selects this option a new modal is shown with the CPU and Memory
usage of the selected Pod. The metrics are updated every 3 seconds.