Add preview widget and rework navigation
It is now possible to show a preview of resource within the details view
of a resource. This can be used to show all Events for a resource or to
show all Pods for a Deployment.

To be able to re-use the existing list and details page we had to remove
the bindings and we have to initialize the controller within the screen.
We also have to set a tag so that a new controller is created every time
the parameters are changed.