Improve logging
With this commit we are introducing a new Logger, which can be used to
write log messages. The Logger takes a "name" which should contain the
class name and method were the log line is written and a "message" and
optional "detail".

With the new Logger can also be used to view the logs within the app, so
that users can now attach the logs to an issue they create in the
repository.