Improve resource usage metrics
We are now using the "fl_chart" package to render the resource usage
charts in the app. We also adjusted the logic for getting the required
metrics, so that we can use the same charts from the home page in the
node details page.

We also fixed the display name of the app for iOS so that it is now
showing "kubenav" instead of "Kubenav".