Minor fixes for iOS and Android
On iOS the backspace key is now working as expected in the terminal, so
that the terminal do not break when a user uses the key to delete a
character.

On Android we had a type error in the native code, which is now fixed.
The problem was within the passed "since" parameter for the logs, which
is an Interger an not a Long value.

On Android we also changed the foreground image for the adaptive icon.
The image now contains the kubenav icon with the circle.