Fix cluster handling
We had a bug for the manual cluster provided, so that it was not
possible to add a cluster, whith the user provided values. This is fixed
now by adding the correct controllers to the text input fields.

It is now also possible to edit clusters, for this we mainly reused the
manual cluster widget, but now we are initializing the values for the
different text input fields with the values of the selected cluster.

Finally we added all the images and configurations for the supported
cloud providers.