Add support for Azure
It is now possible to add new clusters from Azure by providing a
subscription id, tenant id, client id and client secret. We will use
these credentials to get a list of all Kubernetes clusters from the
Azure API. A user can then select the cluster he wants to add from the
returned Azure clusters. The selected cluster are then available within
kubenav.

To be able to add the Azure support we had to introduce the provider
configuration, which will be used for all cloud providers to save the
users configuration for them. It is important to know that the name of a
provider configuration can not be changed, because we have to save a
refrence in the cluster configuration. We also block the deleteion of a
provider configuration as long as the config is used in an existing
cluster.

We also reuse the logic for parsing a kubeconfig file from the Azure
provider for the Kubeconfig provider, so that a user should get more
information why a cluster can not be added from a kubeconfig file.